# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "NatuEve のバックエンド API",
        "title": "NatuEve API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/events": {
            "get": {
                "description": "公開イベントを指定ソート順で返す。認証不要。\nsort は \"created_at\"(デフォルト) / \"event_date\" のみ許可。不正値はデフォルトに戻す。\norder は \"desc\"(デフォルト) / \"asc\" のみ許可。不正値はデフォルトに戻す。\nprifileはProfileSummaryを返す。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "イベント一覧取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ソートカラム(created_at|event_date, default: created_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "ソート順(asc|desc, default: desc)",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "取得件数(default 20, 最大 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "取得開始位置(default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.InternalErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "認証済みユーザーが新規イベントを投稿する。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "イベント投稿",
                "parameters": [
                    {
                        "description": "イベント投稿リクエスト",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.CreateEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.UnauthorizedErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.InternalErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{id}": {
            "get": {
                "description": "指定されたイベントIDの詳細情報を取得する",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "イベント詳細取得",
                "parameters": [
                    {
                        "type": "string",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/events/{id}/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "認証は任意。ログイン時のみ profileId が記録される。\nAuthorization ヘッダなし → 匿名参加（profileId = null）。\nヘッダありでトークンが無効 → 401 で中断。\nヘッダありで有効 → profileId を記録してログイン参加。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "event"
                ],
                "summary": "イベント参加",
                "parameters": [
                    {
                        "type": "string",
                        "description": "イベントID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "参加申込",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.JoinEventRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.JoinEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ValidationErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.UnauthorizedErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.NotFoundErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ConflictErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.InternalErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ConflictErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ConflictErrorBody"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.CreateEventRequest": {
            "description": "イベント投稿に必要な情報。",
            "type": "object",
            "required": [
                "costs",
                "description",
                "eventDate",
                "location",
                "title"
            ],
            "properties": {
                "capacity": {
                    "description": "Capacity は定員（任意・0=未設定・正数=定員）。",
                    "type": "integer",
                    "minimum": 0,
                    "example": 30
                },
                "costs": {
                    "description": "Costs は費用内訳（1件以上必須）。",
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventCostInput"
                    }
                },
                "description": {
                    "description": "Description はイベント説明（必須）。",
                    "type": "string",
                    "example": "春の桜を観察するイベントです。"
                },
                "eventDate": {
                    "description": "EventDate はイベント開催日時(RFC3339)（必須）。",
                    "type": "string",
                    "example": "2026-07-01T10:00:00Z"
                },
                "externalUrl": {
                    "description": "ExternalURL は関連URLs（任意・255文字以内・http/https）。",
                    "type": "string",
                    "maxLength": 255,
                    "example": "https://example.com/event"
                },
                "imageFilenames": {
                    "description": "ImageFilenames は画像の元ファイル名一覧（任意）。指定時は ImageObjectKeys と同数・同順。\nダウンロード時のファイル名（Content-Disposition）と UI 表示に使う。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "imageObjectKeys": {
                    "description": "ImageObjectKeys は画像オブジェクトキーの一覧（任意）。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "items": {
                    "description": "Items は持ち物リスト（任意）。",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventItemInput"
                    }
                },
                "location": {
                    "description": "Location は開催場所（必須・255文字以内）。",
                    "type": "string",
                    "maxLength": 255,
                    "example": "東京都新宿御苑"
                },
                "pdfFilenames": {
                    "description": "PdfFilenames はPDFの元ファイル名一覧（任意）。指定時は PdfObjectKeys と同数・同順。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pdfObjectKeys": {
                    "description": "PdfObjectKeys はPDFオブジェクトキーの一覧（任意・各要素255文字以内）。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "description": "Title はイベントタイトル（必須・255文字以内）。",
                    "type": "string",
                    "maxLength": 255,
                    "example": "サクラ観察会"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.CreateEventResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt はレコード作成日時(RFC3339)。",
                    "type": "string",
                    "example": "2026-06-23T12:00:00Z"
                },
                "id": {
                    "description": "ID は生成されたイベントの UUID。",
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                }
            }
        },
...
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code は機械可読なエラーコード。",
                    "type": "string",
                    "example": "internal_error"
                },
                "message": {
                    "description": "Message は人間向けのエラーメッセージ。",
                    "type": "string",
                    "example": "サーバー内部でエラーが発生しました"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ErrorBody"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventCostInput": {
            "type": "object",
            "required": [
                "category"
            ],
            "properties": {
                "category": {
                    "description": "Category は費用カテゴリ（例: \"参加費\"）。",
                    "type": "string",
                    "maxLength": 255
                },
                "cost": {
                    "description": "Cost は費用（円）。0 以上の整数。",
                    "type": "integer",
                    "minimum": 0
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventCostResponse": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "cost": {
                    "type": "integer"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventItemInput": {
            "type": "object",
            "required": [
                "item"
            ],
            "properties": {
                "isRequired": {
                    "description": "IsRequired は必須かどうか。",
                    "type": "boolean"
                },
                "item": {
                    "description": "Item は持ち物名（例: \"双眼鏡\"）。",
                    "type": "string",
                    "maxLength": 255
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventItemResponse": {
            "type": "object",
            "properties": {
                "isRequired": {
                    "type": "boolean"
                },
                "item": {
                    "type": "string"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventListResponse": {
            "type": "object",
            "properties": {
                "events": {
                    "description": "Events はイベントサマリーの一覧。",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventSummary"
                    }
                },
                "limit": {
                    "description": "Limit は正規化後の実際に使われた取得件数。",
                    "type": "integer",
                    "example": 20
                },
                "offset": {
                    "description": "Offset は正規化後の実際に使われた取得開始位置。",
                    "type": "integer",
                    "example": 0
                },
                "totalCount": {
                    "description": "TotalCount はフィルタなし全件数。クライアントが最終ページ offset を算出するために使う。",
                    "type": "integer",
                    "example": 153
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventResponse": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "costs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventCostResponse"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "externalUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageFilenames": {
                    "description": "ImageFilenames は ImageObjectKeys に対応する元ファイル名（未設定は空文字）。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "imageObjectKeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "imageUrls": {
                    "description": "ImageUrls は ImageObjectKeys に対応する表示用の完全URL。\n公開ベースURL（R2_PUBLIC_BASE_URL）未設定時は空配列。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventItemResponse"
                    }
                },
                "location": {
                    "type": "string"
                },
                "pdfFilenames": {
                    "description": "PdfFilenames は PdfObjectKeys に対応する元ファイル名（未設定は空文字）。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pdfObjectKeys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "pdfUrls": {
                    "description": "PdfUrls は PdfObjectKeys に対応する表示用の完全URL。\n公開ベースURL（R2_PUBLIC_BASE_URL）未設定時は空配列。",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "profile": {
                    "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ProfileSummary"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.EventSummary": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt はレコード作成日時(RFC3339)。",
                    "type": "string",
                    "example": "2026-06-22T12:00:00Z"
                },
                "eventDate": {
                    "description": "EventDate はイベント開催日時(RFC3339)。",
                    "type": "string",
                    "example": "2026-07-01T10:00:00Z"
                },
                "id": {
                    "description": "ID はイベントの一意識別子(UUID)。",
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "location": {
                    "description": "Location は開催場所（文字列）。",
                    "type": "string",
                    "example": "東京都新宿御苑"
                },
                "profile": {
                    "description": "ProfileSummary は投稿者プロフィールのサマリー情報。",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ProfileSummary"
                        }
                    ]
                },
                "profileId": {
                    "description": "ProfileID は投稿者のプロフィール ID(UUID)。",
                    "type": "string",
                    "example": "d290f1ee-6c54-4b01-90e6-d701748f0851"
                },
                "title": {
                    "description": "Title はイベントタイトル。",
                    "type": "string",
                    "example": "サクラ観察会"
                }
            }
        },
...
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.InternalErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code は機械可読なエラーコード。",
                    "type": "string",
                    "example": "internal_error"
                },
                "message": {
                    "description": "Message は人間向けのエラーメッセージ。",
                    "type": "string",
                    "example": "サーバー内部でエラーが発生しました"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.InternalErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.InternalErrorBody"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.JoinEventRequest": {
            "description": "イベント参加申込に必要な情報。認証は任意。",
            "type": "object",
            "required": [
                "mailAddress",
                "username"
            ],
            "properties": {
                "mailAddress": {
                    "description": "MailAddress は参加するユーザーのメールアドレス（必須）。",
                    "type": "string",
                    "maxLength": 255,
                    "example": "yamada@example.com"
                },
                "username": {
                    "description": "Username は参加するユーザーの表示名（必須・255文字以内）。",
                    "type": "string",
                    "maxLength": 255,
                    "example": "山田太郎"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.JoinEventResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "description": "CreatedAt は参加申込日時。",
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "eventId": {
                    "description": "EventID は参加したイベントのUUID。",
                    "type": "string",
                    "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                },
                "mailAddress": {
                    "description": "MailAddress は参加するユーザーのメールアドレス。",
                    "type": "string",
                    "example": "yamada@example.com"
                },
                "profileId": {
                    "description": "ProfileID は参加するユーザーのUUID。ログイン時のみ記録され、匿名参加時は null。",
                    "type": "string",
                    "example": "b2c3d4e5-f6a7-8901-bcde-f23456789012"
                },
                "username": {
                    "description": "Username は参加するユーザーの表示名。",
                    "type": "string",
                    "example": "山田太郎"
                }
            }
        },
...
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.NotFoundErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.NotFoundErrorBody"
                }
            }
        },
...
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.UnauthorizedErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code は機械可読なエラーコード。",
                    "type": "string",
                    "example": "unauthorized"
                },
                "message": {
                    "description": "Message は人間向けのエラーメッセージ。",
                    "type": "string",
                    "example": "認証が必要です"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.UnauthorizedErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.UnauthorizedErrorBody"
                }
            }
        },
...
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ValidationErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code は機械可読なエラーコード。",
                    "type": "string",
                    "example": "invalid_request"
                },
                "message": {
                    "description": "Message は人間向けのエラーメッセージ。",
                    "type": "string",
                    "example": "タイトルは必須です"
                }
            }
        },
        "github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/github_com_GokujyouKaisennDonnburi_NatuEve_API_internal_model.ValidationErrorBody"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Supabase Auth が発行した JWT を \"Bearer \u003ctoken\u003e\" 形式で指定する",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}